requirements.txt
cachetools==5.2.0
future==0.18.2
pefile==2022.5.30
Pillow==9.2.0
py2exe==0.11.1.1
